model_pipeline/voice-backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
sentence-transformers==2.7.0
faiss-cpu
numpy==1.26.4
openai
pandas
tqdm==4.66.1
requests==2.31.0
google-cloud-storage==2.13.0
pytest==7.4.3
pytest-cov==4.1.0
gunicorn==21.2.0
google-auth==2.29.0